still broken, but some stuff fixed
